# apps/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ErrorResponse:
    properties:
      detail:
        description: optional internal error
        example: validation failed on field 'email'
        type: string
      error:
        description: user-friendly message
        example: Invalid request
        type: string
    type: object
...
  user.ChangePasswordRequest:
    properties:
      currentPassword:
        type: string
      newPassword:
        type: string
    type: object
  user.CreateUserRequest:
    properties:
      email:
        type: string
      fullName:
        type: string
      password:
        description: raw password from client
        type: string
      phone:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/user.Role'
        enum:
        - admin
        - driver
        - customer
      slug:
        type: string
    required:
    - email
    - fullName
    - password
    - phone
    - role
    - slug
    type: object
  user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  user.LoginResponse:
    properties:
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      role:
        type: string
      token:
        type: string
    type: object
  user.Role:
    enum:
    - merchant
    - admin
    - driver
    - customer
    - guest
    type: string
    x-enum-varnames:
    - Merchant
    - Admin
    - Driver
    - Customer
    - Guest
  user.UpdateDriverUserProfileRequest:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  user.UpdateUserProfileRequest:
    properties:
      email:
        type: string
      fullName:
        type: string
      phone:
        type: string
    required:
    - email
    - fullName
    - phone
    type: object
  user.UpdateUserRequest:
    properties:
      column:
        type: string
      value: {}
    required:
    - column
    - value
    type: object
  user.UpdateUserStatusRequest:
    properties:
      status:
        $ref: '#/definitions/user.UserStatus'
    required:
    - status
    type: object
  user.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      last_login:
        type: string
      must_change_password:
        type: boolean
      phone:
        type: string
      role:
        $ref: '#/definitions/user.Role'
      slug:
        description: adminSlug used in public route
        type: string
      status:
        $ref: '#/definitions/user.UserStatus'
      updated_at:
        type: string
    type: object
  user.UserStatus:
    enum:
    - active
    - inactive
    - suspended
    - pending
    type: string
    x-enum-varnames:
    - Active
    - Inactive
    - Suspended
    - Pending
host: localhost:8000
info:
  contact: {}
  description: This is the API for fastabiz operations.
  title: FastaBiz API
  version: "1.0"
paths:
...
  /public/create:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email, etc.
      parameters:
      - description: User Input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new user
      tags:
      - public
  /public/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using email and password and returns a JWT
        token.
      parameters:
      - description: User login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.LoginResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Login user
      tags:
      - public
...
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User profile deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
  /users/{id}/driver_profile:
    patch:
      consumes:
      - application/json
      description: Updates only the phone number of a driver (commonly used after
        onboarding)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Driver phone update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.UpdateDriverUserProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user (driver) phone number
      tags:
      - users
  /users/{id}/password:
    put:
      consumes:
      - application/json
      description: Verifies the user's current password and updates it to a new password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Current and new passwords
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid user ID or request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Current password incorrect
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change a user's password
      tags:
      - users
  /users/{id}/profile:
    patch:
      consumes:
      - application/json
      description: Updates the user's name, email, and/or phone number. Partial updates
        allowed.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User profile update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
  /users/{id}/status:
    patch:
      consumes:
      - application/json
      description: Updates the status (e.g., active, inactive, suspended) of a specific
        user.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: User status update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User status updated
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID or request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update a user's status
      tags:
      - users
  /users/{id}/update:
    put:
      consumes:
      - application/json
      description: Updates a user's specific field (e.g., FullName, Email) based on
        user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Field and value to update
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid user ID or request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a specific user field
      tags:
      - users
  /users/all_users:
    get:
      description: Get a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
  /users/by-email/{email}:
    get:
      description: Retrieve a user by their Email
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Invalid Email
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user by Email
      tags:
      - users
  /users/by-id/{id}:
    get:
      description: Retrieve a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"